// packages/core/src/types/view-computed.d.ts
export interface ComputedNode<A extends AnyAux = AnyAux> extends aux.WithTags<A>, aux.WithOptionalLinks {
    id: scalar.NodeId;
    kind: aux.ElementKind<A> | aux.DeploymentKind<A> | '@group';
    parent: scalar.NodeId | null;
    /**
     * Reference to model element
     */
    modelRef?: aux.Fqn<A>;
    /**
     * Reference to deployment element
     */
    deploymentRef?: aux.DeploymentFqn<A>;
    title: string;
    /**
     * Description of the node
     * either summary or description
     */
    description?: scalar.MarkdownOrString | null;
    technology?: string | null;
    children: scalar.NodeId[];
    inEdges: scalar.EdgeId[];
    outEdges: scalar.EdgeId[];
    shape: ElementShape;
    color: Color;
    icon?: Icon | null;
    style: ComputedNodeStyle;
    navigateTo?: aux.StrictViewId<A> | null;
    level: number;
    depth?: number | null;
    /**
     * If this node was customized in the view
     */
    isCustomized?: boolean;
    notation?: string | null;
    notes?: scalar.MarkdownOrString | undefined;
}
export interface ComputedEdge<A extends AnyAux = AnyAux> extends aux.WithOptionalTags<A> {
    id: scalar.EdgeId;
    parent: scalar.NodeId | null;
    source: scalar.NodeId;
    target: scalar.NodeId;
    label: string | null;
    description?: scalar.MarkdownOrString | null;
    technology?: string | null;
    relations: scalar.RelationId[];
    kind?: aux.RelationKind<A> | typeof scalar.StepEdgeKind;
    notation?: string;
    notes?: scalar.MarkdownOrString;
    color: Color;
    line: RelationshipLineType;
    head?: RelationshipArrowType;
    tail?: RelationshipArrowType;
    navigateTo?: aux.StrictViewId<A> | null;
    /**
     * If this edge is derived from custom relationship predicate
     */
    isCustomized?: boolean;
    /**
     * Path to the AST node relative to the view body ast
     * Available only in dynamic views
     * @internal
     */
    astPath?: string;
    /**
     * For layouting purposes
     * @default 'forward'
     */
    dir?: 'forward' | 'back' | 'both';
}
export interface ComputedRankConstraint {
    type: RankValue;
    nodes: scalar.NodeId[];
}
interface BaseComputedViewProperties<A extends AnyAux> extends BaseViewProperties<A>, ViewWithHash, ViewWithNotation {
    readonly [_stage]: 'computed';
    readonly autoLayout: ViewAutoLayout;
    readonly nodes: ReadonlyArray<ComputedNode<A>>;
    readonly edges: ReadonlyArray<ComputedEdge<A>>;
    /**
     * If the view has manual layout (v2)
     */
    readonly hasManualLayout?: boolean;
    /**
     * Manual layout data (v1), for compatibility during migration
     * @deprecated
     */
    readonly manualLayout?: ViewManualLayout | undefined;
}
export interface ComputedElementView<A extends AnyAux = AnyAux> extends BaseComputedViewProperties<A> {
    readonly [_type]: 'element';
    readonly viewOf?: aux.StrictFqn<A>;
    readonly extends?: aux.StrictViewId<A>;
    readonly ranks?: ComputedRankConstraint[];
}
export interface ComputedDeploymentView<A extends AnyAux = AnyAux> extends BaseComputedViewProperties<A> {
    readonly [_type]: 'deployment';
}
export interface ComputedDynamicView<A extends AnyAux = AnyAux> extends BaseComputedViewProperties<A> {
    readonly [_type]: 'dynamic';
    /**
     * How to display the dynamic view
     * - `diagram`: display as a regular likec4 view
     * - `sequence`: display as a sequence diagram
     */
    readonly variant: DynamicViewDisplayVariant;
}